split test suite to speed up CLion
